% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{specvec_methods}
\alias{specvec_methods}
\title{List registered methods}
\usage{
specvec_methods()
}
\value{
A character vector of registered method names.
}
\description{
List registered methods
}
\examples{
specvec_methods()
}
